// src/app/shared/models/transports/dispatchorders/dispatchorder.ts
export interface Dispatchorder {
        id?: number;
        refNo?: string;
        branchId?: number;
        shippingUnitId?: number;
        shippingUnitCode?: string;
        shippingUnitName?: string;
        referCode?: string;
        isSubcontractors?: boolean;
        vihicleId?: number;
        vihiclelLicensePlates?: string;
        vihicleType?: number;
        vehicleTypeCharge?: number;
        vehicleTypeName?:string;
        moocId?: number;
        moocLicensePlates?: string;
        driverId?: number;
        driverName?: string;
        driverTel?: string;
        secondDriverId?:number;
        secondDriverName?: string;
        secondDriverTel?: string;
        fuelDriverId?:number;
        fuelDriverNae?:string;
        numberOfVihicle?: number;
        numberOfPackage?: number;
        weight?: number;
        volume?: number;
        numberOfPallet?: number;
        numberOfCont?: number;
        contType?: string;
        documentType?: string;
        documentNumber?: string;
        fullRoute?: string;
        oilQuota?: number;
        noLoadingOilQuota?:number;
        oilPrice?: number;
        quotaAdjusted?: number;
        quotaAdjustedReason?: string;
        subcontractorsQuoteCode?: string;
        subcontractorsQuoteRouteCode?: string;
        subcontractorsPrice?:number;
        dispatchSummarize?: string;
        inquiryTimeToTheFactory?: Date | string;
        inquiryTimeToThePorts?: Date | string;
        contactInformation?: string;
        note?: string;
        routeId?:number;
        routeName?:string;
        kmRouteQuota?:number;
        kmQuota?: number;
        reasons?:string;
        noLoadingRouteId?:number;
        noLoadingRouteName?:string;
        noLoadingKmRouteQuota?:number;
        noLoadingKmQuota?: number;
        noLoadingReasons?:string;
        shippingTaskId?:number
        shippingType?:number;
        kmOdor?:number;
        startVehicleOdor?: number;
        startEupOdor?: number;
        finishVehicleOdor?: number;
        finishEupOdor?: number;
        startVehicleOdor2?: number;
        finishVehicleOdor2?: number;
        startVehicleOdor3?: number;
        finishVehicleOdor3?: number;
        locked?: boolean;
        finished?: boolean;
        finishedDate?:Date | string;
        noteFinished?:string;
        status?:number;
        rStatus?:string;
        gradeDriver?:number;
        evaluationDriver?:string;
        gradeCs?: number;
        evaluationCs?: string;
        gradeOpMan?: number;
        evaluationOpMan?: string;
        gradeSubcontractors?:number;
        evaluationSubcontractors?:string;
        createdDate?: string;
        createdBy?: string;
        createdByName?:string;
        updatedDate?: Date | string;
        deleted?: string;
        closingDate?:string;
        closingBy?:string;
        listDispatchOrderDetailed?:Dispatchorderdetailed[];
        listDispatchOrderRoutes?:Dispatchorderroutes[];
        listDispatchOrderPurchasePrice?:DispatchOrderPurchasePrice[];
        listDispatchOrderFee?:DispatchOrderFee[];
        listDispatchOrderTicket?:DispatchOrderTicket[];
        listDispatchOrderEtc?:DispatchOrderEtc[];
        listDispatchOrderMonthlyTicket?:DispatchOrderMonthlyticket[];
        listDispatchOrderProcedure?:DispatchOrderProcedure[];
        listDispatchOrderParkingTicket?:DispatchOrderParkingticket[];
        listDispatchOrderSurcharge?:DispatchOrderSurcharge[];
        listDispatchOrderStandardRoute?:DispatchOrderStandardRoute[];
        checked?:boolean;
        totalRows?:number;
        shippingTaskItem?:ShippingTask;
        purchasePrice?:number;
        tollRouteName?:string;
}